requirements.txt
asgiref==3.8.1
Django==5.1.3
django-jazzmin==3.0.1
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
phonenumbers==8.13.51
pillow==11.0.0
python-decouple==3.8
sqlparse==0.5.2
